requirements.txt
django==4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Markdown==3.4.4
django-filter==23.2
djoser==2.2.0
drf-spectacular==0.26.4
django-cryptography==1.1
python-dotenv==1.0.0
psycopg==3.1.9
psycopg-binary==3.1.9
model-bakery==1.11.0
